Update disable-playbar-slider.js - avoid "pause" when user clicks on the playbar
Hi, thanks a lot for this script.
I'm not a dev but with my basic knowledge yesterday I analysed the CPM.js and found out the code to avoid the the "pause" of the project when the user accidently clicks on the playbar. It works.
I hope it could be helpful.